# Don't calculate start timestamp in noop case 